Promote LiteRT runtime client API
